requirements.txt
python-telegram-bot==20.3
python-dotenv>=1.0.0
g4f==0.5.5.7
asyncio>=3.4.3
pillow>=9.0.0
loguru>=0.7.0 